update(exception): 예외 패키지 구조 분리
