+++ src/main/java/nextstep/subway/line/exception/line/LineNotFoundException.java
@@ -1,22 +1,22 @@
-package nextstep.subway.common.exception.section;
+package nextstep.subway.line.exception.line;
 
 import nextstep.subway.common.exception.ServiceException;
 import org.springframework.http.HttpStatus;
 
 /**
  * packageName : nextstep.subway.common.exception
- * fileName : SectionNotCreateException
+ * fileName : LineNotFoundException
  * author : haedoang
  * date : 2021/12/01
- * description : 중복 구간 예외 클래스
+ * description : 노선 없음 Custom Exception
  */
-public class SectionDuplicateException extends ServiceException {
+public class LineNotFoundException extends ServiceException {
     private static final long serialVersionUID = 1L;
 
     public static final HttpStatus status = HttpStatus.BAD_REQUEST;
-    public static final String message = "이미 등록된 구간 입니다.";
+    public static final String message = "노선이 존재하지 않습니다.";
 
-    public SectionDuplicateException() {
+    public LineNotFoundException() {
         super(status, message);
     }
 }
