+++ src/main/java/nextstep/subway/station/exception/StationNotFoundException.java
@@ -1,22 +1,22 @@
-package nextstep.subway.common.exception.member;
+package nextstep.subway.station.exception;
 
 import nextstep.subway.common.exception.ServiceException;
 import org.springframework.http.HttpStatus;
 
 /**
  * packageName : nextstep.subway.common.exception
- * fileName : MemberNotFoundException
+ * fileName : StationNotFoundException
  * author : haedoang
  * date : 2021/12/01
- * description : 회원 미존재 예외 클래스
+ * description : 역 없음 예외 클래스
  */
-public class MemberNotFoundException extends ServiceException {
+public class StationNotFoundException extends ServiceException {
     private static final long serialVersionUID = 1L;
 
     public static final HttpStatus status = HttpStatus.BAD_REQUEST;
-    public static final String message = "회원이 존재하지 않습니다.";
+    public static final String message = "역이 존재하지 않습니다.";
 
-    public MemberNotFoundException() {
+    public StationNotFoundException() {
         super(status, message);
     }
 }
