+++ src/test/java/nextstep/subway/line/domain/LineTest.java
@@ -1,6 +1,6 @@
 package nextstep.subway.line.domain;
 
-import nextstep.subway.common.exception.distance.IllegalDistanceException;
+import nextstep.subway.line.exception.distance.IllegalDistanceException;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.DisplayName;
