+++ src/test/java/nextstep/subway/path/PathAcceptanceTest.java
@@ -4,9 +4,9 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.common.exception.path.PathBeginIsEndException;
-import nextstep.subway.common.exception.path.PathNotFoundException;
-import nextstep.subway.common.exception.station.StationNotFoundException;
+import nextstep.subway.path.exception.PathBeginIsEndException;
+import nextstep.subway.path.exception.PathNotFoundException;
+import nextstep.subway.station.exception.StationNotFoundException;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.path.dto.PathResponse;
